Moved evcent description to own util
